+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 **0.9.20**
 
+* AngularClient - Profile page - Delete project fix
+* AngularClient - Profile page - Media queries fix
 * Package updates
 
 **0.9.19**